+++ beacon-chain/rpc/eth/beacon/blocks_test.go
@@ -905,173 +905,6 @@ func TestServer_GetBlockSSZV2(t *testing.T) {
 	})
 }
 
-func TestServer_GetBlockRoot(t *testing.T) {
-	beaconDB := dbTest.SetupDB(t)
-	ctx := context.Background()
-
-	genBlk, blkContainers := fillDBTestBlocks(ctx, t, beaconDB)
-	headBlock := blkContainers[len(blkContainers)-1]
-
-	t.Run("get root", func(t *testing.T) {
-		wsb, err := blocks.NewSignedBeaconBlock(headBlock.Block.(*ethpbalpha.BeaconBlockContainer_Phase0Block).Phase0Block)
-		require.NoError(t, err)
-		mockChainFetcher := &mock.ChainService{
-			DB:                  beaconDB,
-			Block:               wsb,
-			Root:                headBlock.BlockRoot,
-			FinalizedCheckPoint: &ethpbalpha.Checkpoint{Root: blkContainers[64].BlockRoot},
-			FinalizedRoots:      map[[32]byte]bool{},
-		}
-		bs := &Server{
-			BeaconDB:              beaconDB,
-			ChainInfoFetcher:      mockChainFetcher,
-			HeadFetcher:           mockChainFetcher,
-			OptimisticModeFetcher: mockChainFetcher,
-			FinalizationFetcher:   mockChainFetcher,
-		}
-
-		root, err := genBlk.Block.HashTreeRoot()
-		require.NoError(t, err)
-
-		tests := []struct {
-			name    string
-			blockID []byte
-			want    []byte
-			wantErr bool
-		}{
-			{
-				name:    "bad formatting",
-				blockID: []byte("3bad0"),
-				wantErr: true,
-			},
-			{
-				name:    "canonical slot",
-				blockID: []byte("30"),
-				want:    blkContainers[30].BlockRoot,
-			},
-			{
-				name:    "head",
-				blockID: []byte("head"),
-				want:    headBlock.BlockRoot,
-			},
-			{
-				name:    "finalized",
-				blockID: []byte("finalized"),
-				want:    blkContainers[64].BlockRoot,
-			},
-			{
-				name:    "genesis",
-				blockID: []byte("genesis"),
-				want:    root[:],
-			},
-			{
-				name:    "genesis root",
-				blockID: root[:],
-				want:    root[:],
-			},
-			{
-				name:    "root",
-				blockID: blkContainers[20].BlockRoot,
-				want:    blkContainers[20].BlockRoot,
-			},
-			{
-				name:    "non-existent root",
-				blockID: bytesutil.PadTo([]byte("hi there"), 32),
-				wantErr: true,
-			},
-			{
-				name:    "slot",
-				blockID: []byte("40"),
-				want:    blkContainers[40].BlockRoot,
-			},
-			{
-				name:    "no block",
-				blockID: []byte("105"),
-				wantErr: true,
-			},
-		}
-		for _, tt := range tests {
-			t.Run(tt.name, func(t *testing.T) {
-				blockRootResp, err := bs.GetBlockRoot(ctx, &ethpbv1.BlockRequest{
-					BlockId: tt.blockID,
-				})
-				if tt.wantErr {
-					require.NotEqual(t, err, nil)
-					return
-				}
-				require.NoError(t, err)
-				assert.DeepEqual(t, tt.want, blockRootResp.Data.Root)
-			})
-		}
-	})
-
-	t.Run("execution optimistic", func(t *testing.T) {
-		wsb, err := blocks.NewSignedBeaconBlock(headBlock.Block.(*ethpbalpha.BeaconBlockContainer_Phase0Block).Phase0Block)
-		require.NoError(t, err)
-		mockChainFetcher := &mock.ChainService{
-			DB:                  beaconDB,
-			Block:               wsb,
-			Root:                headBlock.BlockRoot,
-			FinalizedCheckPoint: &ethpbalpha.Checkpoint{Root: blkContainers[64].BlockRoot},
-			Optimistic:          true,
-			FinalizedRoots:      map[[32]byte]bool{},
-			OptimisticRoots: map[[32]byte]bool{
-				bytesutil.ToBytes32(headBlock.BlockRoot): true,
-			},
-		}
-		bs := &Server{
-			BeaconDB:              beaconDB,
-			ChainInfoFetcher:      mockChainFetcher,
-			HeadFetcher:           mockChainFetcher,
-			OptimisticModeFetcher: mockChainFetcher,
-			FinalizationFetcher:   mockChainFetcher,
-		}
-		blockRootResp, err := bs.GetBlockRoot(ctx, &ethpbv1.BlockRequest{
-			BlockId: []byte("head"),
-		})
-		require.NoError(t, err)
-		assert.Equal(t, true, blockRootResp.ExecutionOptimistic)
-	})
-
-	t.Run("finalized", func(t *testing.T) {
-		wsb, err := blocks.NewSignedBeaconBlock(headBlock.Block.(*ethpbalpha.BeaconBlockContainer_Phase0Block).Phase0Block)
-		require.NoError(t, err)
-		mockChainFetcher := &mock.ChainService{
-			DB:                  beaconDB,
-			Block:               wsb,
-			Root:                headBlock.BlockRoot,
-			FinalizedCheckPoint: &ethpbalpha.Checkpoint{Root: blkContainers[64].BlockRoot},
-			Optimistic:          true,
-			FinalizedRoots: map[[32]byte]bool{
-				bytesutil.ToBytes32(blkContainers[32].BlockRoot): true,
-				bytesutil.ToBytes32(blkContainers[64].BlockRoot): false,
-			},
-		}
-		bs := &Server{
-			BeaconDB:              beaconDB,
-			ChainInfoFetcher:      mockChainFetcher,
-			HeadFetcher:           mockChainFetcher,
-			OptimisticModeFetcher: mockChainFetcher,
-			FinalizationFetcher:   mockChainFetcher,
-		}
-
-		t.Run("true", func(t *testing.T) {
-			blockRootResp, err := bs.GetBlockRoot(ctx, &ethpbv1.BlockRequest{
-				BlockId: []byte("32"),
-			})
-			require.NoError(t, err)
-			assert.Equal(t, true, blockRootResp.Finalized)
-		})
-		t.Run("false", func(t *testing.T) {
-			blockRootResp, err := bs.GetBlockRoot(ctx, &ethpbv1.BlockRequest{
-				BlockId: []byte("64"),
-			})
-			require.NoError(t, err)
-			assert.Equal(t, false, blockRootResp.Finalized)
-		})
-	})
-}
-
 func TestServer_ListBlockAttestations(t *testing.T) {
 	ctx := context.Background()
