+++ beacon-chain/rpc/eth/beacon/handlers.go
@@ -7,18 +7,23 @@ import (
 	"fmt"
 	"io"
 	"net/http"
+	"strconv"
 
+	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/go-playground/validator/v10"
 	"github.com/pkg/errors"
 	"github.com/prysmaticlabs/prysm/v4/beacon-chain/core/transition"
 	"github.com/prysmaticlabs/prysm/v4/beacon-chain/rpc/eth/shared"
 	"github.com/prysmaticlabs/prysm/v4/consensus-types/blocks"
 	"github.com/prysmaticlabs/prysm/v4/consensus-types/interfaces"
+	"github.com/prysmaticlabs/prysm/v4/consensus-types/primitives"
+	"github.com/prysmaticlabs/prysm/v4/encoding/bytesutil"
 	http2 "github.com/prysmaticlabs/prysm/v4/network/http"
 	ethpbv1 "github.com/prysmaticlabs/prysm/v4/proto/eth/v1"
 	ethpbv2 "github.com/prysmaticlabs/prysm/v4/proto/eth/v2"
 	"github.com/prysmaticlabs/prysm/v4/proto/migration"
 	eth "github.com/prysmaticlabs/prysm/v4/proto/prysm/v1alpha1"
+	"go.opencensus.io/trace"
 )
 
 const (
@@ -630,4 +635,116 @@ func (bs *Server) validateEquivocation(blk interfaces.ReadOnlyBeaconBlock) error
 		return fmt.Errorf("block for slot %d already exists in fork choice", blk.Slot())
 	}
 	return nil
+}
+
+// GetBlockRoot retrieves hashTreeRoot of ReadOnlyBeaconBlock/BeaconBlockHeader.
+func (bs *Server) GetBlockRoot(w http.ResponseWriter, r *http.Request) {
+	//(*ethpbv1.BlockRootResponse, error)
+	ctx, span := trace.StartSpan(r.Context(), "beacon.GetBlockRoot")
+	defer span.End()
+
+	blockID := r.URL.Query().Get("block_id")
+	valid := shared.ValidateHex(w, "Block ID", blockID)
+	if !valid {
+		return
+	}
+	blockIDBytes, err := hexutil.Decode(blockID)
+	if err != nil {
+		http2.HandleError(w, "Could not decode block ID into bytes: "+err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	var root []byte
+	switch blockID {
+	case "head":
+		root, err = bs.ChainInfoFetcher.HeadRoot(ctx)
+		if err != nil {
+			http2.HandleError(w, "Could not retrieve head block: "+err.Error(), http.StatusInternalServerError)
+			return
+		}
+		if root == nil {
+			http2.HandleError(w, "No head root was found", http.StatusNotFound)
+			return
+		}
+	case "finalized":
+		finalized := bs.ChainInfoFetcher.FinalizedCheckpt()
+		root = finalized.Root
+	case "genesis":
+		blk, err := bs.BeaconDB.GenesisBlock(ctx)
+		if err != nil {
+			http2.HandleError(w, "Could not retrieve blocks for genesis slot: "+err.Error(), http.StatusInternalServerError)
+			return
+		}
+		if err := blocks.BeaconBlockIsNil(blk); err != nil {
+			http2.HandleError(w, "Could not find genesis block: "+err.Error(), http.StatusNotFound)
+			return
+		}
+		blkRoot, err := blk.Block().HashTreeRoot()
+		if err != nil {
+			http2.HandleError(w, "Could not hash genesis block", http.StatusInternalServerError)
+			return
+		}
+		root = blkRoot[:]
+	default:
+		if len(blockID) == 32 {
+			var blockID32 [32]byte
+			copy(blockID32[:], blockIDBytes)
+			blk, err := bs.BeaconDB.Block(ctx, blockID32)
+			if err != nil {
+				http2.HandleError(w, fmt.Sprintf("Could not retrieve block for block root %#x: %v", blockID, err), http.StatusInternalServerError)
+				return
+			}
+			if err := blocks.BeaconBlockIsNil(blk); err != nil {
+				http2.HandleError(w, "Could not find block: "+err.Error(), http.StatusNotFound)
+				return
+			}
+			root = blockIDBytes
+		} else {
+			slot, err := strconv.ParseUint(string(blockID), 10, 64)
+			if err != nil {
+				http2.HandleError(w, "Could not parse block ID: "+err.Error(), http.StatusBadRequest)
+				return
+			}
+			hasRoots, roots, err := bs.BeaconDB.BlockRootsBySlot(ctx, primitives.Slot(slot))
+			if err != nil {
+				http2.HandleError(w, fmt.Sprintf("Could not retrieve blocks for slot %d: %v", slot, err), http.StatusInternalServerError)
+				return
+			}
+
+			if !hasRoots {
+				http2.HandleError(w, "Could not find any blocks with given slot", http.StatusNotFound)
+				return
+			}
+			root = roots[0][:]
+			if len(roots) == 1 {
+				break
+			}
+			for _, blockRoot := range roots {
+				canonical, err := bs.ChainInfoFetcher.IsCanonical(ctx, blockRoot)
+				if err != nil {
+					http2.HandleError(w, "Could not determine if block root is canonical: "+err.Error(), http.StatusInternalServerError)
+					return
+				}
+				if canonical {
+					root = blockRoot[:]
+					break
+				}
+			}
+		}
+	}
+
+	b32Root := bytesutil.ToBytes32(root)
+	isOptimistic, err := bs.OptimisticModeFetcher.IsOptimisticForRoot(ctx, b32Root)
+	if err != nil {
+		http2.HandleError(w, "Could not check if block is optimistic: "+err.Error(), http.StatusInternalServerError)
+		return
+	}
+	response := &ethpbv1.BlockRootResponse{
+		Data: &ethpbv1.BlockRootContainer{
+			Root: root,
+		},
+		ExecutionOptimistic: isOptimistic,
+		Finalized:           bs.FinalizationFetcher.IsFinalized(ctx, b32Root),
+	}
+	http2.WriteJson(w, response)
 }
