+++ proto/eth/service/beacon_chain_service.proto
@@ -230,15 +230,6 @@ service BeaconChain {
     };
   }
 
-  // GetBlockRoot retrieves hashTreeRoot of BeaconBlock/BeaconBlockHeader.
-  //
-  // Spec: https://ethereum.github.io/beacon-APIs/?urls.primaryName=v2.3.0#/Beacon/getBlockRoot
-  rpc GetBlockRoot(v1.BlockRequest) returns (v1.BlockRootResponse) {
-    option (google.api.http) = {
-      get: "/internal/eth/v1/beacon/blocks/{block_id}/root"
-    };
-  }
-
   // GetBlock retrieves block details for given block id.
   //
   // Spec: https://ethereum.github.io/beacon-APIs/?urls.primaryName=v2.3.0#/Beacon/getBlock
